Revert "Change underlying duration type from double to long double"
This reverts commit 2a159bd6fa6d726c819058d6c07cf132f22c4a52.

Conflicts:
	amx_profiler/statistics_writer_json.cpp

+++ amx_profiler/duration.h
@@ -119,14 +119,14 @@ private:
 	ValueType count_;
 };
 
-typedef Duration<long double, Ratio<1, 1000000000> > Nanoseconds;
-typedef Duration<long double, Ratio<1, 1000000> >    Microseconds;
-typedef Duration<long double, Ratio<1, 1000> >       Milliseconds;
-typedef Duration<long double, Ratio<1, 1> >          Seconds;
-typedef Duration<long double, Ratio<60, 1> >         Minutes;
-typedef Duration<long double, Ratio<3600, 1> >       Hours;
-typedef Duration<long double, Ratio<86400, 1> >      Days;
-typedef Duration<long double, Ratio<604800, 1> >     Weeks;
+typedef Duration<double, Ratio<1, 1000000000> > Nanoseconds;
+typedef Duration<double, Ratio<1, 1000000> >    Microseconds;
+typedef Duration<double, Ratio<1, 1000> >       Milliseconds;
+typedef Duration<double, Ratio<1, 1> >          Seconds;
+typedef Duration<double, Ratio<60, 1> >         Minutes;
+typedef Duration<double, Ratio<3600, 1> >       Hours;
+typedef Duration<double, Ratio<86400, 1> >      Days;
+typedef Duration<double, Ratio<604800, 1> >     Weeks;
 
 } // namespace amx_profiler
